+++ src/main/java/com/teragrep/cfe_41/configs/cfe_04/global/Global.java
@@ -43,28 +43,37 @@
  * Teragrep, the applicable Commercial License may apply to this file if you as
  * a licensee so wish it.
  */
-package com.teragrep.cfe_41.configs.cfe_04;
+package com.teragrep.cfe_41.configs.cfe_04.global;
 
+import com.teragrep.cfe_41.configs.cfe_04.Jsonable;
 import jakarta.json.Json;
 import jakarta.json.JsonStructure;
-import java.util.Map;
+
 import java.util.Objects;
 
-public final class Meta implements Jsonable {
+public final class Global implements Jsonable {
 
-    private final String name;
+    private final String truncate;
+    private final String lci;
+    private final String lciMalformed;
+    private final String maxDaysAgo;
 
-    public Meta(final Map<String, String> config) {
-        this(config.get("cfe_04.meta.name"));
-    }
-
-    public Meta(final String name) {
-        this.name = name;
+    public Global(final String truncate, final String lci, final String lciMalformed, final String maxDaysAgo) {
+        this.truncate = truncate;
+        this.lci = lci;
+        this.lciMalformed = lciMalformed;
+        this.maxDaysAgo = maxDaysAgo;
     }
 
     @Override
     public JsonStructure asJsonStructure() {
-        return Json.createObjectBuilder().add("name", name).build();
+        return Json
+                .createObjectBuilder()
+                .add("truncate", truncate)
+                .add("last_chance_index", lci)
+                .add("last_chance_index_malformed", lciMalformed)
+                .add("max_days_ago", maxDaysAgo)
+                .build();
     }
 
     @Override
@@ -72,12 +81,13 @@ public final class Meta implements Jsonable {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        final Meta meta = (Meta) o;
-        return Objects.equals(name, meta.name);
+        final Global global = (Global) o;
+        return Objects.equals(truncate, global.truncate) && Objects.equals(lci, global.lci)
+                && Objects.equals(lciMalformed, global.lciMalformed) && Objects.equals(maxDaysAgo, global.maxDaysAgo);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hashCode(name);
+        return Objects.hash(truncate, lci, lciMalformed, maxDaysAgo);
     }
 }
