+++ src/main/java/com/teragrep/cfe_41/configs/cfe_04/index/IndexFactory.java
@@ -1,4 +1,4 @@
-package com.teragrep.cfe_41.configs.cfe_04;
+package com.teragrep.cfe_41.configs.cfe_04.index;
 
 import java.util.Map;
