+++ src/main/java/com/teragrep/cfe_41/configs/cfe_04/sourcetype/SourcetypeFactory.java
@@ -1,4 +1,4 @@
-package com.teragrep.cfe_41.configs.cfe_04;
+package com.teragrep.cfe_41.configs.cfe_04.sourcetype;
 
 import java.util.Map;
 import java.util.List;
