+++ src/test/java/com/teragrep/cfe_41/configs/cfe_04/GlobalTest.java
@@ -45,6 +45,7 @@
  */
 package com.teragrep.cfe_41.configs.cfe_04;
 
+import com.teragrep.cfe_41.configs.cfe_04.global.Global;
 import jakarta.json.Json;
 import jakarta.json.JsonObject;
 import nl.jqno.equalsverifier.EqualsVerifier;
